+++ sdk/spring/azure-spring-boot/src/test/java/com/azure/spring/autoconfigure/jms/NonPremiumServiceBusJMSAutoConfigurationTest.java
@@ -3,8 +3,11 @@
 
 package com.azure.spring.autoconfigure.jms;
 
+import java.util.logging.Logger;
 import org.apache.qpid.jms.JmsConnectionFactory;
+import org.apache.qpid.jms.policy.JmsDefaultPrefetchPolicy;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.autoconfigure.jms.JmsAutoConfiguration;
@@ -12,8 +15,16 @@ import org.springframework.boot.test.context.FilteredClassLoader;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class NonPremiumServiceBusJMSAutoConfigurationTest extends AbstractServiceBusJMSAutoConfigurationTest {
+
+    private static final Logger LOGGER = Logger.getLogger(NonPremiumServiceBusJMSAutoConfigurationTest.class.getName());
+
+    @BeforeAll
+    public static void init() {
+        LOGGER.info("Starting NonPremiumServiceBusJMSAutoConfigurationTest");
+    }
 
     @Test
     public void testAzureServiceBusNonPremiumAutoConfiguration() {
@@ -35,6 +46,25 @@ public class NonPremiumServiceBusJMSAutoConfigurationTest extends AbstractServic
         contextRunner.withClassLoader(new FilteredClassLoader(JmsConnectionFactory.class))
                      .run(context -> assertThat(context).doesNotHaveBean(AzureServiceBusJMSProperties.class));
     }
+
+    @Test
+    public void testNonPreminumPrefetchPolicy() {
+        ApplicationContextRunner contextRunner = getContextRunnerWithProperties();
+        contextRunner.run(
+            context -> {
+                assertThat(context).hasBean("jmsConnectionFactory");
+                JmsConnectionFactory jmsConnectionFactory = (JmsConnectionFactory) context.getBean("jmsConnectionFactory");
+                assertNotNull(jmsConnectionFactory.getPrefetchPolicy());
+                assertThat(jmsConnectionFactory.getPrefetchPolicy() instanceof JmsDefaultPrefetchPolicy).isTrue();
+                JmsDefaultPrefetchPolicy policy = (JmsDefaultPrefetchPolicy) jmsConnectionFactory.getPrefetchPolicy();
+                assertThat(policy.getTopicPrefetch()).isEqualTo(12);
+                assertThat(policy.getDurableTopicPrefetch()).isEqualTo(999);
+                assertThat(policy.getQueuePrefetch()).isEqualTo(19);
+                assertThat(policy.getQueueBrowserPrefetch()).isEqualTo(21);
+            }
+        );
+    }
+
 
     @Test
     public void testAzureServiceBusJMSPropertiesPricingTireValidation() {
@@ -74,7 +104,12 @@ public class NonPremiumServiceBusJMSAutoConfigurationTest extends AbstractServic
                 "spring.jms.servicebus.idle-timeout=123",
                 "spring.jms.servicebus.pricing-tier=basic",
                 "spring.jms.servicebus.listener.reply-pub-sub-domain=false",
-                "spring.jms.servicebus.listener.reply-qos-settings.priority=1"
+                "spring.jms.servicebus.listener.reply-qos-settings.priority=1",
+                "spring.jms.servicebus.prefetch-policy.all=5",
+                "spring.jms.servicebus.prefetch-policy.topic-prefetch=12",
+                "spring.jms.servicebus.prefetch-policy.durable-topic-prefetch=999",
+                "spring.jms.servicebus.prefetch-policy.queue-prefetch= 19",
+                "spring.jms.servicebus.prefetch-policy.queue-browser-prefetch= 21"
             );
     }
 }